Create a package using charpente and usethis conventions

A charpente powered package does not differ from other package structure. create_charpente calls a series of function to help you quickly setting up the package structure.

Usage

create_charpente(path, remote = NULL, private = FALSE, license)

Arguments

path

A path. If it exists, it is used. If it does not exist, it is created, provided that the parent path exists.

remote

Name of the remote Github organization to connect to.

private

Whether the repository is private. Default to FALSE.

license

Which license is your project under?

Examples

if (FALSE) { # \dontrun{
 create_charpente("bs4Dash")
} # }